Added prompts merge
